fix: split worker consumption rates to satisfy economy file-size gate
Keep the Labour gist helpers public without growing economy_consumption_phases.dart past the phase-8 260-line ceiling.
